requirements.txt
torch
opencv-python
pillow
lmdb
PyYAML
albumentations
tensorboard
torchvision